+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0
+
+### Added
+
+- `Sidekiq::TransactionGuard.testing` now automatically sets the allowed transaction level when the block begins. This provides better support transactional fixtures in test environments.
+- Added `Sidekiq::TransactionGuard.disable` method to allow temporarily disabling the transaction guard within a block. This is useful in test environments when you want to setup data for your tests without worrying about transaction levels.
+- Added `count` parameter to `set_allowed_transaction_level` to allow setting the allowed transaction level explicitly. This is useful for test setups where the transaction level cannot be determined automatically, such as when using ActiveRecord transactional fixtures.
+- Added Railtie for automatic integration with Rails applications.
+- Added helpers for easier testing setup with RSpec.
+- Added `Sidekiq::TransactionGuard::Middleware.init` method to simplify middleware initialization.
+- Added minitest helper module for easier integration with Minitest test suites.
+
+### Removed
+
+- Removed support for ActiveRecord versions prior to 6.0.
+- Removed support for Sidekiq versions prior to 6.0.
+
 ## 1.0.3
 
 ### Changed